Fix detection of Blade Symphony on Windows x64 (#67)
* Fix detection of Blade Symphony on Windows x64

Port finding the module base address on x64 builds from SourceMod.
Fix a bunch of truncation warnings on x64 while at it too.

* Fix mm_TrimComments stripping last char before comment

```c
char test[] = "hi this is a a test!// comment";
mm_TrimComments(test);
puts(test); // prints "hi this is a test" without the !
```

* Fix unsigned comparison warnings

* Fix truncation warnings in SourceHook::String

* Fix unsigned comparison warnings

clang does have some nice analysis.

+++ core/metamod.h
@@ -72,7 +72,7 @@ public:
 	void GetShVersions(int &shvers, int &shimpl);
 	void AddListener(ISmmPlugin *plugin, IMetamodListener *pListener);
 	void *MetaFactory(const char *iface, int *ret, PluginId *id);
-	int FormatIface(char iface[], unsigned int maxlength);
+	int FormatIface(char iface[], size_t maxlength);
 	void *InterfaceSearch(CreateInterfaceFn fn, const char *iface, int max, int *ret);
 	const char *GetBaseDir();
 	size_t PathFormat(char *buffer, size_t len, const char *fmt, ...);
